Updated TextAPI SDK

+++ TextMood.Shared/Models/ITextMoodModel.cs
@@ -6,7 +6,7 @@ namespace TextMood.Shared
     {
 		string Id { get; set; }
 		string Text { get; set; }
-        float? SentimentScore { get; set; }
+        double? SentimentScore { get; set; }
 		DateTimeOffset CreatedAt { get; set; }
 		DateTimeOffset UpdatedAt { get; set; }
 		bool IsDeleted { get; set; }
