+++ TextMood/Models/TextMoodModel.cs
@@ -15,7 +15,7 @@ namespace TextMood
 
 		public string Id { get; set; }
 		public string Text { get; set; }
-		public float? SentimentScore { get; set; }
+		public double? SentimentScore { get; set; }
 		public DateTimeOffset CreatedAt { get; set; }
 		public DateTimeOffset UpdatedAt { get; set; }
 		public bool IsDeleted { get; set; }
